refactor: consolidate DataStore logic into AuthenticationLocalDataSourceImpl
This commit refactors the DataStore implementation by:

- Deleting the `AppPreferences` interface and `AuthenticationLocalDataSourceImpl` (the one implementing `AppPreferences`).
- Renaming `AppDataStorePreferences` to `AuthenticationLocalDataSourceImpl`.
- Making the new `AuthenticationLocalDataSourceImpl` implement `AuthenticationLocalSource` directly.
- Updating DI and tests to reflect these changes.

This simplifies the DataStore interaction by removing an unnecessary abstraction layer and directly using the `AuthenticationLocalDataSourceImpl` for managing session data.
