+++ localDatasource/src/main/java/com/amsterdam/localdatasource/dataStore/AuthenticationLocalDataSourceImpl.kt
@@ -1,19 +1,20 @@
-package com.amsterdam.localdatasource.dataStore.appPreferences
+package com.amsterdam.localdatasource.dataStore
 
 import androidx.datastore.core.DataStore
 import androidx.datastore.preferences.core.Preferences
 import androidx.datastore.preferences.core.edit
 import androidx.datastore.preferences.core.stringPreferencesKey
 import com.amsterdam.domain.exceptions.UnauthorizedException
-import com.amsterdam.localdatasource.dataStore.appPreferences.AppDataStorePreferences.PreferenceKeys.SESSION_ID
-import com.amsterdam.localdatasource.dataStore.appPreferences.AppDataStorePreferences.PreferenceKeys.SESSION_TYPE
-import com.amsterdam.localdatasource.dataStore.datasource.AppPreferences
+import com.amsterdam.localdatasource.dataStore.AuthenticationLocalDataSourceImpl.PreferenceKeys.SESSION_ID
+import com.amsterdam.localdatasource.dataStore.AuthenticationLocalDataSourceImpl.PreferenceKeys.SESSION_TYPE
+import com.amsterdam.repository.datasource.local.AuthenticationLocalSource
 import kotlinx.coroutines.flow.first
+import kotlinx.coroutines.flow.firstOrNull
 import kotlinx.coroutines.flow.map
 
-class AppDataStorePreferences(
+class AuthenticationLocalDataSourceImpl(
     private val datastore: DataStore<Preferences>,
-) : AppPreferences {
+) : AuthenticationLocalSource {
     private object PreferenceKeys {
         val SESSION_TYPE = stringPreferencesKey("sessionType")
         val SESSION_ID = stringPreferencesKey("session_id")
@@ -37,14 +38,15 @@ class AppDataStorePreferences(
         }
     }
 
-    override suspend fun getSessionId(): String? {
+    override suspend fun getCachedSessionId(): String {
         return datastore.data
             .map { setting ->
                 setting[SESSION_ID] ?: throw UnauthorizedException()
-            }.first()
+            }.firstOrNull()
+            ?: throw UnauthorizedException()
     }
 
-    override suspend fun clearSessionId() {
+    override suspend fun clearCachedSessionId() {
         datastore.edit { setting ->
             setting[SESSION_ID] = ""
         }
