--- localDatasource/src/main/java/com/amsterdam/localdatasource/dataStore/datasource/AppPreferences.kt
@@ -1,12 +1,0 @@
-package com.amsterdam.localdatasource.dataStore.datasource
-
-interface AppPreferences {
-    suspend fun getSessionType(): String
-    suspend fun setSessionType(sessionType: String)
-
-    suspend fun cacheSessionId(sessionId: String)
-
-    suspend fun getSessionId(): String?
-
-    suspend fun clearSessionId()
-}
