--- localDatasource/src/main/java/com/amsterdam/localdatasource/dataStore/datasource/AuthenticationLocalDataSourceImpl.kt
@@ -1,26 +1,0 @@
-package com.amsterdam.localdatasource.dataStore.datasource
-
-import com.amsterdam.domain.exceptions.UnauthorizedException
-import com.amsterdam.repository.datasource.local.AuthenticationLocalSource
-
-class AuthenticationLocalDataSourceImpl(
-    private val appPreferences: AppPreferences
-): AuthenticationLocalSource {
-    override suspend fun setSessionType(sessionType: String) {
-        appPreferences.setSessionType(sessionType)
-    }
-
-    override suspend fun getSessionType(): String {
-        return appPreferences.getSessionType()
-    }
-
-    override suspend fun cacheSessionId(sessionId: String) {
-        appPreferences.cacheSessionId(sessionId = sessionId)
-    }
-
-    override suspend fun getCachedSessionId(): String = appPreferences.getSessionId() ?: throw UnauthorizedException()
-
-    override suspend fun clearCachedSessionId() {
-        appPreferences.clearSessionId()
-    }
-}
