+++ localDatasource/src/test/java/com/amsterdam/localdatasource/dataStore/appPreferences/AppDataStorePreferencesTest.kt
@@ -3,6 +3,7 @@ package com.amsterdam.localdatasource.dataStore.appPreferences
 import androidx.datastore.core.DataStore
 import androidx.datastore.preferences.core.PreferenceDataStoreFactory
 import androidx.datastore.preferences.core.Preferences
+import com.amsterdam.localdatasource.dataStore.AuthenticationLocalDataSourceImpl
 import com.google.common.truth.Truth.assertThat
 import kotlinx.coroutines.test.runTest
 import org.junit.jupiter.api.AfterEach
@@ -10,17 +11,17 @@ import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import java.io.File
 
-class AppDataStorePreferencesTest {
+class AuthenticationLocalDataSourceImplTest {
     private lateinit var dataStore: DataStore<Preferences>
     private lateinit var file: File
-    private lateinit var appDataStorePreferences: AppDataStorePreferences
+    private lateinit var authenticationLocalDataSourceImpl: AuthenticationLocalDataSourceImpl
     @BeforeEach
     fun setup() {
         file = File.createTempFile("test", ".preferences_pb").apply {
             deleteOnExit()
         }
         dataStore = PreferenceDataStoreFactory.create { file }
-        appDataStorePreferences = AppDataStorePreferences(dataStore)
+        authenticationLocalDataSourceImpl = AuthenticationLocalDataSourceImpl(dataStore)
     }
 
     @AfterEach
@@ -31,7 +32,7 @@ class AppDataStorePreferencesTest {
     @Test
     fun `getSessionType should return empty string when no value is set`() = runTest {
         //When
-        val result = appDataStorePreferences.getSessionType()
+        val result = authenticationLocalDataSourceImpl.getSessionType()
         //Then
         assertThat(result).isEmpty()
     }
@@ -41,9 +42,9 @@ class AppDataStorePreferencesTest {
         //Given
         val userType = "guest"
         //When
-        appDataStorePreferences.setSessionType(userType)
+        authenticationLocalDataSourceImpl.setSessionType(userType)
         //Then
-        val result = appDataStorePreferences.getSessionType()
+        val result = authenticationLocalDataSourceImpl.getSessionType()
         assertThat(result).isEqualTo(userType)
     }
 }
